+++ src/routers/router.js
@@ -20,4 +20,6 @@ router.get('/logout', userAuthController.logoutAndRedirect);
 
 
 
+
+
 module.exports = router;
